cloud_functions/pubsub/requirements.txt
google-cloud-pubsub
google-cloud-secret-manager
functions-framework
gunicorn